por fin me salio csm

+++ src/controllers/image.js
@@ -10,26 +10,38 @@ ctrl.index = (req, res) => {
 
 };
 
-ctrl.create = async (req, res) => {
-    const imgUrl = randomNumber();
-    console.log(imgUrl);
-    const imageTempPath = req.file.path;
-    const ext = path.extname(req.file.originalname).toLowerCase();
-    const targetPath = path.resolve('src/public/upload/${imgUrl}${ext}')
+ctrl.create = (req, res) => {
 
-    if (ext === '.png' || ext === '.jpg' || ext === '.jpeg' || ext === '.gif') {
-        await fs.rename(imageTempPath, targetPath);
-        const newImg = new Image({
-            title: req.body.title,
-            filename: imgUrl + ext,
-            description: req.body.description
-        });
-        const imageSaved = await newImg.save();
-    } else {
-        await fs.unlink(imageTempPath);
-        res.status(500).json({error: 'Only Images are allowed'});
-    }
-    res.send('works!');
+    const saveImage = async () => {
+        const imgUrl = randomNumber();
+        const images = await Image.find({filename: imgUrl});
+        if (images.length > 0) {
+            saveImage();
+        } else {
+                console.log(imgUrl);
+            const imageTempPath = req.file.path;
+            const ext = path.extname(req.file.originalname).toLowerCase();
+            const targetPath = path.resolve(`src/public/upload/${imgUrl}${ext}`)
+
+            if (ext === '.png' || ext === '.jpg' || ext === '.jpeg' || ext === '.gif') {
+                await fs.rename(imageTempPath, targetPath);
+                const newImg = new Image({
+                    title: req.body.title,
+                    filename: imgUrl + ext,
+                    description: req.body.description
+                });
+                const imageSaved = await newImg.save();
+                //res.redirect('/images/:image_id');
+                res.send('Works!');
+            } else {
+                await fs.unlink(imageTempPath);
+                res.status(500).json({error: 'Only Images are allowed'});
+            }
+        }
+    };
+
+    saveImage();
+
 };
 
 ctrl.like = (req, res) => {
